refactor: split credential deploy routes
